Cast: build fix for win 64
BUG=307582

Review URL: https://codereview.chromium.org/27171004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@228791 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/audio_receiver/audio_decoder_unittest.cc
@@ -46,6 +46,7 @@ TEST_F(AudioDecoderTest, Pcm16MonoNoResampleOnePacket) {
   uint8* payload_data = reinterpret_cast<uint8*>(&payload[0]);
   size_t payload_size = payload.size() * sizeof(int16);
 
+  // TODO(pwestin): change API to size_t to avoid casts.
   audio_decoder_->IncomingParsedRtpPacket(payload_data,
       static_cast<int>(payload_size), rtp_header);
 
@@ -93,6 +94,7 @@ TEST_F(AudioDecoderTest, Pcm16StereoNoResampleTwoPackets) {
   uint8* payload_data = reinterpret_cast<uint8*>(&payload[0]);
   size_t payload_size = payload.size() * sizeof(int16);
 
+  // TODO(pwestin): change API to size_t to avoid casts.
   audio_decoder_->IncomingParsedRtpPacket(payload_data,
       static_cast<int>(payload_size), rtp_header);
 
@@ -118,8 +120,10 @@ TEST_F(AudioDecoderTest, Pcm16StereoNoResampleTwoPackets) {
 
   rtp_header.webrtc.header.sequenceNumber++;
   rtp_header.webrtc.header.timestamp += (audio_config.frequency / 100) * 2 * 2;
-  audio_decoder_->IncomingParsedRtpPacket(payload_data, payload_size,
-                                          rtp_header);
+
+  // TODO(pwestin): change API to size_t to avoid casts.
+  audio_decoder_->IncomingParsedRtpPacket(payload_data,
+      static_cast<int>(payload_size), rtp_header);
 
   EXPECT_TRUE(audio_decoder_->GetRawAudioFrame(number_of_10ms_blocks,
                                                desired_frequency,
@@ -158,6 +162,7 @@ TEST_F(AudioDecoderTest, Pcm16Resample) {
   uint8* payload_data = reinterpret_cast<uint8*>(&payload[0]);
   size_t payload_size = payload.size() * sizeof(int16);
 
+  // TODO(pwestin): change API to size_t to avoid casts.
   audio_decoder_->IncomingParsedRtpPacket(payload_data,
       static_cast<int>(payload_size), rtp_header);
